app/requirements.txt
azure-core==1.31.0
azure-storage-blob==12.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
pydantic==2.6.4
sqlalchemy==2.0.40
flask_sqlalchemy==3.0.3
sqlalchemy_utils==0.37.8
pydantic[email]
pymysql
icecream